+++ test/selenium/test_data.py
@@ -1,4 +1,10 @@
 
+from datetime import datetime
+import json
+
+from terrareg.database import Database
+
+
 integration_git_providers = {
     1: {
         'name': 'testgitprovider',
@@ -103,7 +109,7 @@ integration_test_data = {
                 }
             },
             'datadog': {
-                'id': 57,
+                'id': 59,
                 'versions': {
                     '1.0.0': {'published': True}
                 }
@@ -355,13 +361,200 @@ integration_test_data = {
         'fullypopulated': {'testprovider': {
             'id': 56,
             'versions': {
+                # Older version
+                '1.2.0': {'published': True},
+                # Newer unpublished version
+                '1.6.0': {},
+                # Newer published beta version
+                '1.6.1-beta': {'published': True, 'beta': True},
+                # Unpublished and beta version
+                '1.0.0-beta': {'published': False, 'beta': True},
                 '1.5.0': {
                     'description': 'This is a test module version for tests.',
                     'owner': 'This is the owner of the module',
                     'repo_base_url_template': 'https://link-to.com/source-code-here',
-                    'published': True
+                    'published': True,
+                    'beta': False,
+                    'internal': False,
+                    'published_at': datetime(2022, 1, 5, 22, 53, 12),
+                    'readme_content': '# This is an exaple README!',
+                    'variable_template': json.dumps([
+                        {
+                            'name': 'name_of_application',
+                            'type': 'string',
+                            'quote_value': True,
+                            'additional_help': 'Provide the name of the application'
+                        }
+
+                    ]),
+                    'module_details': json.dumps({
+                        'header': '',
+                        'footer': '',
+                        'inputs': [
+                            {
+                                'name': 'name_of_application',
+                                'type': 'string',
+                                'description': 'Enter the application name',
+                                'default': None,
+                                'required': True
+                            },
+                            {
+                                'name': 'string_with_default_value',
+                                'type': 'string',
+                                'description': 'Override the default string',
+                                'default': 'this is the default',
+                                'required': False
+                            },
+                            {
+                                'name': 'example_boolean_input',
+                                'type': 'bool',
+                                'description': 'Override the truthful boolean',
+                                'default': True,
+                                'required': False
+                            },
+                            {
+                                'name': 'example_list_input',
+                                'type': 'list',
+                                'description': 'Override the stringy list',
+                                'default': ['value 1', 'value 2'],
+                                'required': False
+                            }
+                        ],
+                        'modules': [],
+                        'outputs': [
+                            {
+                                'name': 'generated_name',
+                                'description': 'Name with randomness'
+                            },
+                            {
+                                'name': 'no_desc_output',
+                                'description': None
+                            }
+                        ],
+                        'providers': [
+                            {
+                                'name': 'random',
+                                'alias': 'random-alias',
+                                'version': '5.2.1'
+                            },
+                            {
+                                'name': 'someothercompany/unsafe',
+                                'alias': None,
+                                'version': '2.0.0'
+                            }
+                        ],
+                        'requirements': [],
+                        'resources': [
+                            {
+                                'type': 'string',
+                                'name': 'random_suffix',
+                                'provider': 'random',
+                                'source': 'hashicorp/random',
+                                'mode': 'managed',
+                                'version': 'latest',
+                                'description': None
+                            }
+                        ]
+                    }),
+                    'examples': {
+                        'examples/test-example': {
+                            'example_files': {
+                                'examples/test-example/data.tf': '# This contains data objects',
+                                'examples/test-example/variables.tf': 'variable "test" {\n  description = "test variable"\n  type = string\n}',
+                                'examples/test-example/main.tf': '# Call root module\nmodule "root" {\n  source = "../../"\n}'
+                            },
+                            'readme_content': '# Example 1 README',
+                            'module_details': json.dumps({
+                                'header': '',
+                                'footer': '',
+                                'inputs': [
+                                    {
+                                        'name': 'input_for_example',
+                                        'type': 'string',
+                                        'description': 'Enter the example name',
+                                        'default': None,
+                                        'required': True
+                                    }
+                                ],
+                                'modules': [],
+                                'outputs': [
+                                    {
+                                        'name': 'example_output',
+                                        'description': 'Example name with randomness'
+                                    }
+                                ],
+                                'providers': [
+                                    {
+                                        'name': 'example_random',
+                                        'alias': None,
+                                        'version': None
+                                    }
+                                ],
+                                'requirements': [],
+                                'resources': [
+                                    {
+                                        'type': 'string',
+                                        'name': 'example_random_suffix',
+                                        'provider': 'example_random',
+                                        'source': 'hashicorp/example_random',
+                                        'mode': 'managed',
+                                        'version': 'latest',
+                                        'description': None
+                                    }
+                                ]
+                            })
+                        }
+                    },
+                    'submodules': {
+                        'modules/example-submodule1': {
+                            'readme_content': '# Submodule 1 README',
+                            'module_details': json.dumps({
+                                'header': '',
+                                'footer': '',
+                                'inputs': [
+                                    {
+                                        'name': 'input_for_submodule',
+                                        'type': 'string',
+                                        'description': 'Enter the submodule name',
+                                        'default': None,
+                                        'required': True
+                                    }
+                                ],
+                                'modules': [],
+                                'outputs': [
+                                    {
+                                        'name': 'submodule_output',
+                                        'description': 'Submodule name with randomness'
+                                    }
+                                ],
+                                'providers': [
+                                    {
+                                        'name': 'submodule_random',
+                                        'alias': None,
+                                        'version': None
+                                    }
+                                ],
+                                'requirements': [],
+                                'resources': [
+                                    {
+                                        'type': 'string',
+                                        'name': 'submodule_random_suffix',
+                                        'provider': 'submodule_random',
+                                        'source': 'hashicorp/submodule_random',
+                                        'mode': 'managed',
+                                        'version': 'latest',
+                                        'description': None
+                                    }
+                                ]
+                            })
+                        }
+                    }
                 },
             }
+        }},
+        'noversion': {'testprovider': {
+            'id': 57,
+            'versions': {}
         }}
     },